+++ lib/src/runner/remote_listener.dart
@@ -98,8 +98,14 @@ class RemoteListener {
               : OperatingSystem.find(message['os']),
           path: message['path']);
 
-      Invoker
-          .guard(() => new RemoteListener._(suite, printZone)._listen(channel));
+      runZoned(() {
+        Invoker.guard(
+            () => new RemoteListener._(suite, printZone)._listen(channel));
+      },
+          // Make the declarer visible to running tests so that they'll throw
+          // useful errors when calling `test()` and `group()` within a test,
+          // and so they can add to the declarer's `tearDownAll()` list.
+          zoneValues: {#test.declarer: declarer});
     }, onError: (error, stackTrace) {
       _sendError(channel, error, stackTrace, verboseChain);
     }, zoneSpecification: new ZoneSpecification(print: (_, __, ___, line) {